requirements.txt
numpy==2.2.5
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.2.3
Pillow==11.2.1
pytest==8.3.5
scikit_learn==1.6.1
torch==2.7.0+cu128
torchvision==0.22.0
tqdm==4.67.1
ultralytics==8.3.73
